+++ haystack/preview/document_stores/_utils.py
@@ -1,0 +1,10 @@
+class StoreError(Exception):
+    pass
+
+
+class DuplicateDocumentError(StoreError):
+    pass
+
+
+class MissingDocumentError(StoreError):
+    pass
