--- haystack/v2/data/__init__.py
@@ -1,5 +1,0 @@
-from haystack.v2.data.data import Data, TextData, TableData, ImageData, AudioData
-from haystack.v2.data.document import Document, TextDocument, TableDocument, ImageDocument, AudioDocument
-from haystack.v2.data.query import Query, TextQuery, TableQuery, ImageQuery, AudioQuery
-from haystack.v2.data.answer import Answer, TextAnswer, TableAnswer, ImageAnswer, AudioAnswer
-from haystack.v2.data.span import Span
