--- haystack/v2/stores/_utils.py
@@ -1,10 +1,0 @@
-class StoreError(Exception):
-    pass
-
-
-class DuplicateError(StoreError):
-    pass
-
-
-class MissingItemError(StoreError):
-    pass
